% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_propensity_score.R
\name{calculate_power_li_2025}
\alias{calculate_power_li_2025}
\title{Calculate Power Using Li et al. (2025) Method}
\usage{
calculate_power_li_2025(
  n,
  effect_size,
  alpha = 0.05,
  treatment_prop,
  overlap_phi,
  rho_squared,
  weight_type = "ATE",
  outcome_var = 1
)
}
\arguments{
\item{n}{Sample size}

\item{effect_size}{Effect size}

\item{alpha}{Significance level}

\item{treatment_prop}{Treatment prevalence (0-1)}

\item{overlap_phi}{Bhattacharyya overlap coefficient (0-1)}

\item{rho_squared}{Confounder-outcome association (R-squared, 0-1)}

\item{weight_type}{Weighting method}

\item{outcome_var}{Outcome variance}
}
\value{
Statistical power (0-1)
}
\description{
Calculates statistical power for a given sample size using Li et al. (2025) method.
}
