% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_text.R
\name{create_continuous_power_result_text}
\alias{create_continuous_power_result_text}
\title{Create Continuous Outcomes Power Analysis Result Text}
\usage{
create_continuous_power_result_text(n1, n2, d, power, alpha, sided)
}
\arguments{
\item{n1}{Sample size for group 1}

\item{n2}{Sample size for group 2}

\item{d}{Effect size (Cohen's d)}

\item{power}{Calculated power (as proportion)}

\item{alpha}{Significance level}

\item{sided}{Test type ("two.sided", "greater", "less")}
}
\value{
HTML formatted text
}
\description{
Generates formatted result text for continuous outcomes power calculations.
}
