% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui_result_cards.R
\name{create_copy_button}
\alias{create_copy_button}
\title{Create a copy-to-clipboard button}
\usage{
create_copy_button(target_id, button_text = "Copy to Clipboard", class = NULL)
}
\arguments{
\item{target_id}{ID of the element to copy content from}

\item{button_text}{Button label (default: "Copy to Clipboard")}

\item{class}{Additional CSS classes}
}
\value{
HTML button element
}
\description{
Create a copy-to-clipboard button
}
\examples{
create_copy_button("result_text", "Copy Results")

}
