% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui_documentation.R
\name{create_doc_card}
\alias{create_doc_card}
\title{Create Individual Documentation Card}
\usage{
create_doc_card(
  icon,
  title,
  description,
  link,
  link_text = "Learn More",
  new_window = TRUE
)
}
\arguments{
\item{icon}{Icon object (from shiny::icon())}

\item{title}{Card title}

\item{description}{Card description}

\item{link}{Link URL or anchor}

\item{link_text}{Link button text}

\item{new_window}{Open link in new window (default: TRUE for vignettes)}
}
\value{
HTML div element containing a documentation card
}
\description{
Create Individual Documentation Card
}
