% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui_documentation.R
\name{create_documentation_page}
\alias{create_documentation_page}
\title{Create Documentation Page UI}
\usage{
create_documentation_page()
}
\value{
HTML div element containing documentation page
}
\description{
Generates the documentation center with cards for different topics
}
