% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui_help.R
\name{create_global_help}
\alias{create_global_help}
\title{Create global help content (Regulatory Guidance & Interpretation Guide)
To be used in a Help modal or separate page}
\usage{
create_global_help()
}
\description{
Create global help content (Regulatory Guidance & Interpretation Guide)
To be used in a Help modal or separate page
}
