% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_text.R
\name{create_result_header}
\alias{create_result_header}
\title{Create Standard Result Header}
\usage{
create_result_header()
}
\value{
HTML tags for result header
}
\description{
Generates the standard header shown at the top of all result outputs.
}
