% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_text.R
\name{create_survival_equiv_samplesize_text}
\alias{create_survival_equiv_samplesize_text}
\title{Create Time-to-Event Equivalence Sample Size Result Text}
\usage{
create_survival_equiv_samplesize_text(
  n_total,
  n_test,
  n_ref,
  d_events,
  hr_expected,
  hr_lower,
  hr_upper,
  power,
  alpha,
  prop_exposed,
  event_rate
)
}
\arguments{
\item{n_total}{Total sample size required}

\item{n_test}{Sample size for test group}

\item{n_ref}{Sample size for reference group}

\item{d_events}{Number of events required}

\item{hr_expected}{Expected hazard ratio}

\item{hr_lower}{Lower equivalence margin}

\item{hr_upper}{Upper equivalence margin}

\item{power}{Desired power (as proportion)}

\item{alpha}{Significance level}

\item{prop_exposed}{Proportion in exposed/treatment group (as proportion)}

\item{event_rate}{Overall event rate (as proportion)}
}
\value{
HTML formatted text
}
\description{
Generates formatted result text for time-to-event equivalence sample size calculations.
}
