% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_survival_ni.R
\name{events_survival_ni}
\alias{events_survival_ni}
\title{Calculate Number of Events Needed}
\usage{
events_survival_ni(power, hr_expected, hr_margin, alpha = 0.025)
}
\arguments{
\item{power}{Desired power}

\item{hr_expected}{Expected HR}

\item{hr_margin}{NI margin}

\item{alpha}{Alpha level}
}
\value{
Number of events required
}
\description{
Calculates the required number of events for a time-to-event NI/equivalence test.
}
