% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_survival_equivalence.R
\name{fct_survival_equivalence}
\alias{fct_survival_equivalence}
\title{Survival Equivalence/Non-Inferiority Functions}
\description{
Business logic and validation for time-to-event equivalence and non-inferiority testing
}
