% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_text.R
\name{format_cohens_d}
\alias{format_cohens_d}
\title{Create Cohen's d Interpretation}
\usage{
format_cohens_d(d)
}
\arguments{
\item{d}{Cohen's d value}
}
\value{
HTML formatted text
}
\description{
Formats Cohen's d interpretation for continuous outcomes.
}
