% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_text.R
\name{format_hazard_ratio}
\alias{format_hazard_ratio}
\title{Create Hazard Ratio Interpretation}
\usage{
format_hazard_ratio(hr)
}
\arguments{
\item{hr}{Hazard ratio value}
}
\value{
HTML formatted text
}
\description{
Formats hazard ratio interpretation for survival analysis.
}
