% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_text.R
\name{format_minimal_detectable_effect}
\alias{format_minimal_detectable_effect}
\title{Format Minimal Detectable Effect}
\usage{
format_minimal_detectable_effect(p1, p2, effect_measures, h)
}
\arguments{
\item{p1}{Proportion in group 1}

\item{p2}{Proportion in group 2}

\item{effect_measures}{List with RR, OR, RD}

\item{h}{Cohen's h value}
}
\value{
HTML formatted text
}
\description{
Creates a formatted box showing minimal detectable effect measures
}
