% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_survival_ni.R
\name{interpret_hr_margin}
\alias{interpret_hr_margin}
\title{Interpret Hazard Ratio Margin}
\usage{
interpret_hr_margin(hr_margin, test_type = "non-inferiority")
}
\arguments{
\item{hr_margin}{Hazard ratio margin}

\item{test_type}{"non-inferiority" or "equivalence"}
}
\value{
HTML formatted interpretation text
}
\description{
Provides interpretation of a hazard ratio margin for non-inferiority/equivalence testing.
}
