% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_propensity_score.R
\name{interpret_overlap_coefficient}
\alias{interpret_overlap_coefficient}
\title{Interpret Overlap Coefficient (Bhattacharyya phi)}
\usage{
interpret_overlap_coefficient(phi)
}
\arguments{
\item{phi}{Overlap coefficient (0-1)}
}
\value{
List with interpretation details
}
\description{
Provides interpretation of overlap coefficient value
}
