% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_mediation.R
\name{PATH_COEF_NEGLIGIBLE}
\alias{PATH_COEF_NEGLIGIBLE}
\title{Mediation Analysis Power Calculation Helper Functions}
\usage{
PATH_COEF_NEGLIGIBLE
}
\description{
Functions for calculating power and sample size for mediation analysis
}
