% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_export.R
\name{supports_pdf_export}
\alias{supports_pdf_export}
\title{Check if Analysis Type Supports PDF Export}
\usage{
supports_pdf_export(analysis_type)
}
\arguments{
\item{analysis_type}{String identifier for the analysis module}
}
\value{
Logical indicating if PDF export is supported
}
\description{
Determines whether a given analysis type has PDF export functionality
implemented.
}
\details{
Currently only "power_single" and "ss_single" have PDF export templates.
This function can be used to conditionally show/hide PDF export buttons
or display appropriate notifications.
}
