% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_survival_ni.R
\name{survival_ni}
\alias{survival_ni}
\title{Statistical Functions for Time-to-Event Equivalence/Non-Inferiority Testing}
\description{
Functions for sample size and power calculations for time-to-event equivalence
and non-inferiority studies using hazard ratio-based methods.
}
