% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_validation.R
\name{validate_integer_input}
\alias{validate_integer_input}
\title{Validate Integer Input}
\usage{
validate_integer_input(value, name, min = NULL, max = NULL, allow_null = FALSE)
}
\arguments{
\item{value}{Input value to validate}

\item{name}{Parameter name for error messages}

\item{min}{Minimum allowed value (optional)}

\item{max}{Maximum allowed value (optional)}

\item{allow_null}{Whether NULL is acceptable (default: FALSE)}
}
\value{
Validated integer value
}
\description{
Validates and coerces to integer. Useful for count variables
like number of tests, number of clusters, etc.
}
\examples{
\dontrun{
n_tests <- validate_integer_input(input$n_tests, "Number of tests", min = 1)
n_clusters <- validate_integer_input(input$n_clusters, "Number of clusters", min = 2)
}
}
