% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui_result_cards.R
\name{create_recommendations_panel}
\alias{create_recommendations_panel}
\title{Create a recommendations panel}
\usage{
create_recommendations_panel(recommendations, title = "Recommendations")
}
\arguments{
\item{recommendations}{Character vector of recommendation texts}

\item{title}{Panel title (default: "Recommendations")}
}
\value{
HTML div with recommendations list
}
\description{
Create a recommendations panel
}
\examples{
create_recommendations_panel(c(
  "Your ICC (0.05) is typical for clinical outcomes",
  "Consider a discontinuation rate of 0.20 for 12-month studies"
))

}
