% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_propensity_score.R
\name{interpret_rho_squared}
\alias{interpret_rho_squared}
\title{Interpret Confounder-Outcome Association (ρ²)}
\usage{
interpret_rho_squared(rho_squared)
}
\arguments{
\item{rho_squared}{R-squared value (0-1)}
}
\value{
List with interpretation details
}
\description{
Provides interpretation of R-squared representing confounding strength
}
