% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_mediation.R
\docType{data}
\name{PATH_COEF_NEGLIGIBLE}
\alias{PATH_COEF_NEGLIGIBLE}
\title{Mediation Analysis Power Calculation Helper Functions}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
PATH_COEF_NEGLIGIBLE
}
\description{
Functions for calculating power and sample size for mediation analysis
}
\keyword{datasets}
